% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruleset_construction.R
\name{bind_rules}
\alias{bind_rules}
\title{Programatically Combine a List of Rules and Rulesets into a Single Ruleset}
\usage{
bind_rules(rule_ruleset_list)
}
\arguments{
\item{rule_ruleset_list}{a list of rules and rulesets you whish to combine
into a single list}
}
\value{
a ruleset which consolidates all the inputs
}
\description{
Programatically Combine a List of Rules and Rulesets into a Single Ruleset
}
