% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{detect_backend}
\alias{detect_backend}
\title{Detects the backend which will be used for checking the rules}
\usage{
detect_backend(x)
}
\arguments{
\item{x}{The data object, ie a data.frame, tibble, data.table, arrow, or DBI object}
}
\value{
a single character element with the name of the backend to use.
One of \code{base-r}, \code{data.table}, \code{dplyr}, \code{collectibles} (for arrow or DBI objects)
}
\description{
The detection will be made based on the class of the object as well as the packages installed.
For example, if a \code{data.frame} is used, it will look if \code{data.table} or \code{dplyr} are installed on the system, as they provide more speed.
Note the main functions will revert the
}
\examples{
data <- mtcars
detect_backend(data)
}
\seealso{
\code{\link[=check_data]{check_data()}}
}
