% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruleset_construction.R
\name{+.ruleset}
\alias{+.ruleset}
\alias{\%+\%}
\title{Add Rulesets Together}
\usage{
\method{+}{ruleset}(a, b)

a \%+\% b
}
\arguments{
\item{a}{the second ruleset you wish to add}
}
\description{
\code{+} allows you to concatenate several rulesets into a larger rulesets. This
can be useful if you want to create a ruleset for a dataset out of checks
for other datasets.
}
