% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im-export.R
\name{write_rules}
\alias{write_rules}
\alias{read_rules}
\title{Read and write rules to a yaml file}
\usage{
write_rules(x, file)

read_rules(file)
}
\arguments{
\item{x}{a list of rules}

\item{file}{a filename}
}
\value{
the filename invisibly
}
\description{
Read and write rules to a yaml file
}
\section{Functions}{
\itemize{
\item \code{read_rules}: reads a ruleset back in
}}

\examples{
rr <- list(
  rule(mpg > 10),
  rule(cyl \%in\% c(4, 6, 8))
)
file <- tempfile(fileext = ".yml")
write_rules(rr, file)
}
