% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describes a dataset}
\usage{
describe(x)
}
\arguments{
\item{x}{a dataset, either a \code{\link{data.frame}}, \code{\link[dplyr:reexports]{dplyr::tibble}}, \code{\link[data.table:data.table]{data.table::data.table}},
\code{\link[arrow:Table]{arrow::arrow_table}}, \code{\link[arrow:open_dataset]{arrow::open_dataset}}, or \code{\link[dplyr:tbl]{dplyr::tbl}} (SQL connection)}
}
\value{
a \code{data.frame}, \code{dplyr::tibble}, or \code{data.table::data.table} containing
a summary of the dataset given
}
\description{
Note that the current version is in the beta stadium at best, that means the
R-native formats (data.frame, dplyr/tibble, or data.table) are a lot faster
than arrow or SQL-based datasets.
}
\examples{
describe(mtcars)
}
\seealso{
Similar to \href{https://cran.r-project.org/web/packages/skimr/vignettes/skimr.html}{skimr::skim()},
\href{https://cran.r-project.org/web/packages/summarytools/vignettes/introduction.html#data-frame-summaries-dfsummary}{summarytools::dfSummary()},
and \href{https://jthomasmock.github.io/gtExtras/reference/gt_plt_summary.html}{gtExtras::gt_plt_summary()}
}
