% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{ruleset}
\alias{ruleset}
\alias{print.ruleset}
\title{Creates a set of rules}
\usage{
ruleset(...)

\method{print}{ruleset}(x, n = 3, ...)
}
\arguments{
\item{...}{a list of rules}

\item{x}{a ruleset to print}

\item{n}{a maximum number of rules to print}
}
\value{
the list of rules as a ruleset
}
\description{
Creates a set of rules
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Prints a ruleset
}}

\examples{
r1 <- rule(mpg > 10)
r2 <- rule(mpg < 20)
rs <- rule_set(r1, r2)
rs

rs <- ruleset(
  rule(cyl \%in\% c(4, 6, 8)),
  rule(is.numeric(disp))
)
rs
}
