% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruleset_construction.R
\name{datavarifyr_plus}
\alias{datavarifyr_plus}
\title{Add Rules and Rulesets Together}
\usage{
datavarifyr_plus(a, b)
}
\arguments{
\item{a}{the second ruleset you wish to add}
}
\description{
\code{+} allows you to add rules and rulesets into larger rulesets. This
can be useful if you want to create a ruleset for a dataset out of checks
for other datasets.
}
